Allow jobs to be added concurrently
Prior to this commit, jobs were added to a single file, called
CACHE_FILE inside the cache_dir. In particular, every time a job was
added, litani would read the file, append the job, and write the file
back out. This meant that jobs could not be added concurrently.

This commit makes litani write each job to a single file in a directory.
When the user finally runs 'litani run-build', all those files are read
and collated into the CACHE_FILE.