Exit with zero even when pipelines fail by default
This commit changes Litani's default behaviour such that the return code
does not depend on the pipelines' success. Litani will exit with a
return code of zero even when one or more ninja jobs fail. The old
behaviour, where a single ninja job failing would make litani exit with
1, can be restored with `--abnormal-exit-on-fail`.

This change is due to litani having an out-of-band medium for
communicating the status of pipelines (the result.json file), which is
vastly more expressive than a return code. Hence, the return code is
reserved for communicating fatal errors, like a crash or other bug in
litani itself.