Touch 'expired' file for old HTML directories
After symlinking the 'html' directory to a new directory where the HTML
report has recently been written, Litani will now touch a file called
'expired' in the old HTML directory. This is to allow for easy
identification of directories that can be garbage-collected.

As of this commit, it is not safe to garbage-collect expired report
directories if there is a possibility that some other process is still
reading the old directory. This will be addressed in future commits that
implement locking of the report directories.