Add table of contents to pipeline pages
Each pipeline page now displays a table of contents at the top, with one
line for each job. Each line is a hyperlink to the full information box
for that job, as well as having a status marker on the left to quickly
indicate whether the job has passed, failed, or is still in progress.