Support adding jobs to a job pool
It is now possible to specify a list of job pools, each pool having a
'depth', when running `litani init` using the `--pools` flag.
Individual jobs can then be placed in a pool using the `--pool` flag to
`litani add-job`.

If a pool has a depth D, then only D jobs that are part of that pool
will run in parallel at any given moment, even if the overall
parallelism limit is higher than D. This allows limiting the number of
resource-hungry jobs that are running at once.