Add litani dump-run command
This is to give litani runs the ability to 'introspect', that is, to
read the run.json file while the run is still in progress. This means
that clients do not have to know how to locate the cache directory, and
also means that we can guarantee that clients will see an up-to-date
version of the run. That is, if a Litani job runs litani dump-run, the
JSON file that gets printed will show all of the job's
reverse-dependencies has having completed. This is exercised in the e2e
test for this feature.