Add --phony-outputs flag
This commit introduces a new flag to indicate that one or more of a
job's output files are 'phony', meaning that the job will not actually
write the files upon completion.

By default, Litani warns the user when a job has not written one of its
output files, because this can indicate that the build is incorrect. If
a job does not write its output files, then all dependent jobs will run
unconditionally, making the build unnecessarily slow. However, sometimes
users use outputs merely as a way to create a dependency between two
jobs, even when the first job doesn't actually write the output file. In
this case, the --phony-outputs flag can be used to suppress the warning,
so that only genuinely missing output files generate a warning.

This flag means that users no longer need to write 'touchfiles' upon job
completion to generate a dependency ordering. Instead, they can create
build graphs whose input and output nodes are entirely phony, enabling
the creation of dependency chains without having to manage temporary
files.

This commit fixes #29.