bug: Make atomic_write process-safe
This commit fixes the following bug: two litani processes, running in
parallel, both call lib.litani.atomic_write() on the same file. Both
processes will attempt to write to the same temporary file, and then
rename that temporary file to the destination file. This means that (1)
the file might be garbled from concurrent writes and (2) the second
process will crash when it attempts to rename the file, as the first
process will already have renamed it.

This bug specifically manifested in builds where jobs would run `litani
dump-run` in parallel.