Make dump-run more resilient to parallel runs
Previously, if there was a problem reading a dumped run-file, the
dump-run command would wait for one second and try again. Since this
command is inherently racy until we give it a much more sophisticated
implementation (it would really need IPC with the litani run-build
process), this commit should at least make it more reliable when running
it in parallel. The dump-run process now sleeps for a variable amount of
time before attempting to re-load the file; the duration both increases
exponentially and has small random jitters added to it.