Add consistency check if dump-run called from job
This commit makes `litani dump-run` perform a consistency check on the
run before printing it out if the `dump-run` program is called from a
Litani job.

Litani jobs that call `litani dump-run` typically do so because they
want to examine the outcome of reverse-dependencies. That is, if a
Litani job J with inputs I1, I2, I3, ... runs `litani dump-run`, then J
would expect any job whose outputs includes one of I1, I2, I3 to have
the `complete` field set to true. However, due to a possible data race
when running this command, this might not actually be the case. This
commit ensures that `litani dump-run` will continue retrying to print
the run until the reverse dependencies of the calling job are labelled
as complete.