Add success/complete counts to dashboard CI stages
The HTML dashboard now annotates each CI stage with the number of
completed jobs, across the whole run, that are part of that stage, as
well as the number of jobs within the stage that succeeded.

This allows users to quickly count the number of stages (as opposed to
jobs or pipelines) that failed. Since there is no concept of an
'in-progress stage', a successful run should always display equal
numbers of complete and successful stages, even if some of the jobs
within the stage have not yet completed.

This fixes #50.