Create process groups for spawned subprocesses
This commit changes how litani-exec spawns and subsequently kills
subprocesses. The subshell that runs commands is now put into its own
session.

Previously, Litani would send SIGTERM and SIGKILL directly to the
subshell that it spawned to run a single command. This would sometimes
not work in cases where the command ignored these signals. The subshell
would terminate and the underlying command would continue running, being
reparented to init(1).

This commit ensures that the spawned subshell gets put into a new
session, implying that it gets its own process group that all descendent
processes are a member of. This allows litani-exec to retrieve the
process group ID, and send the TERM and KILL signals to the entire
process group.

This fixes a bug where the --timeout argument to litani-add-job(1) would
have no effect when running commands that ignored TERM.