Do not search for litani cache in subdirectories
Prior to this commit, if one were to run Litani in a directory after
first running it inside a subdirectory, unexpected behaviour would be
observed. This is because in each new Litani build, a search was
conducted for the cache pointer (.litani_cache_dir file) across both
parent directories of the cwd and subdirectories eventually causing
build in parent directory to fail.

This commit fixes #66