Show running jobs on pipeline page
Distinguish the jobs that are currently running from the jobs that have
not started on the pipeline HTML page. Use CSS to color jobs that are
currently running white and the jobs that have not started gray on the
dependency graph, border them in black on the pipeline page, and only
show the in progress icon on running jobs in the table of contents.