Acquire report directory before breaking symlink
This commit fixes the code so that newly-written HTML report directories
are acquired before the 'html' symlink is updated to point to them. This
is conceptually the correct order to do this in, although there is no
actual change in the implementation. This commit also adds a comment
that explains the sequence of symlink breaking in the code.