Fix chapter 5 man page descriptions
This change makes the name and description of the Chapter 5 man pages
consistent with the other man pages, i.e. the name of the man page,
followed by a hyphen surrounded by whitespace, followed by the
description.

This is to allow us to parse out the name and description from all man
pages to generate a table of contents.