Do not create new Unix session for subprocesses
This fixes a bug where subprocesses would not be terminated upon
termination of `litani run-build`.

Prior to this commit, every Litani job would run its command in a fresh
Unix session. This disconnected the command from the controlling
terminal, such that a signal delivered from the terminal would not reach
the command. Thus, users who sent INT to `litani run-build` by pressing
Ctrl-C would observe that `litani run-build`, `ninja`, and all of the
`litani exec` invocations would terminate, but the job commands would be
reparented to init(1) and continue to run with no controlling terminal.
This commit ensures that job commands once again receive the INT, TERM,
and KILL signals.

This reverts commit 241c5e8b9a3eed58beeb9f4b7aef1d4bd7b9418b.