Correctly handle signals TERM, INT, and HUP
`litani run-build` now makes ninja the leader of its own process group.
This allows run-build to forward signals to that entire process group,
so that ninja and all its descendants receive the signal. This ensures
that when users press Ctrl-C on Litani's controlling terminal, the INT
is propagated to the commands in each Litani job.

This fixes #162.