Add test fixture for abnormal run-build RCs
The regression test suite can now check the return code of the `litani
run-build` invocation. If a test defines a get_run_build_return_code
method, it will be called to find what the return code is expected to
be. This is in preparation for a future commit that will test that
`litani run-build` correctly exits abnormally.