Modify run-build so artifacts can be uploaded to specified S3 bucket
During an upload, both "incremental snapshots" and a final version
of the report directory are being uploaded. The location in S3 starts
with "BuildArtifacts". It is followed by a subfolder whose name indicates
the build's run ID. There are 2 more nested subfolders distinguishing
the final report directory from the possibly several incremental ones.